Refactor views lessons side panel view handling
