+++ kolibri/plugins/coach/assets/src/routes/lessonsRoutes.js
@@ -133,7 +133,7 @@ export default [
       },
       {
         name: PageNames.LESSON_SELECT_RESOURCES,
-        path: 'select-resources/:topicId?',
+        path: 'select-resources/:viewId?',
         component: LessonResourceSelection,
       },
       {
