+++ kolibri/plugins/coach/assets/src/views/lessons/LessonSummaryPage/sidePanels/LessonResourceSelection/subPages/SelectFromBookmarks.vue
@@ -1,0 +1,14 @@
+<template>
+
+  <div>Select from Bookmarks</div>
+
+</template>
+
+
+<script>
+
+  export default {
+    name: 'SelectFromBookmarks',
+  };
+
+</script>
