+++ kolibri/plugins/coach/assets/src/views/lessons/LessonSummaryPage/sidePanels/LessonResourceSelection/subPages/SelectFromChannels.vue
@@ -1,0 +1,14 @@
+<template>
+
+  <div>Select from Channels</div>
+
+</template>
+
+
+<script>
+
+  export default {
+    name: 'SelectFromChannels',
+  };
+
+</script>
